backend/requirements.txt
fastapi==0.128.5
uvicorn[standard]==0.40.0
python-multipart==0.0.22
numpy==2.2.6
opencv-python-headless==4.10.0.84
paddleocr==3.4.0
paddlepaddle==3.2.2
python-dotenv==1.2.1